etl_oracle_csv/requirements.txt
oracledb==2.5.0
pandas==2.2.3
pyyaml==6.0.2
pytest==8.3.5python-dotenv
